# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateBookingRequest {
  tableId: string;
  hallName: string;
  tableName: string;
  customerName: string;
  customerPhone: string;
  /** Date in YYYY-MM-DD format */
  reservationDate: string;
  /** Time slot e.g. "07:00 PM" */
  reservationTime: string;
  /** @minimum 1 */
  guestCount: number;
  specialRequest?: string | null;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  tableId: string;
  hallName: string;
  tableName: string;
  customerName: string;
  customerPhone: string;
  reservationDate: string;
  reservationTime: string;
  guestCount: number;
  specialRequest?: string | null;
  status: BookingStatus;
  createdAt: string;
}

export type UpdateBookingStatusRequestStatus =
  (typeof UpdateBookingStatusRequestStatus)[keyof typeof UpdateBookingStatusRequestStatus];

export const UpdateBookingStatusRequestStatus = {
  confirmed: "confirmed",
  cancelled: "cancelled",
} as const;

export interface UpdateBookingStatusRequest {
  status: UpdateBookingStatusRequestStatus;
}

export interface BookingError {
  error: string;
  detail?: string;
}

export interface TableStatusMap {
  [key: string]: "available" | "reserved" | "occupied";
}

export type TableStatusEntryStatus =
  (typeof TableStatusEntryStatus)[keyof typeof TableStatusEntryStatus];

export const TableStatusEntryStatus = {
  available: "available",
  occupied: "occupied",
} as const;

export interface TableStatusEntry {
  tableId: string;
  status: TableStatusEntryStatus;
}

export type UpdateTableStatusRequestStatus =
  (typeof UpdateTableStatusRequestStatus)[keyof typeof UpdateTableStatusRequestStatus];

export const UpdateTableStatusRequestStatus = {
  available: "available",
  occupied: "occupied",
} as const;

export interface UpdateTableStatusRequest {
  status: UpdateTableStatusRequestStatus;
}

export type BulkUpdateStatusRequestStatuses = {
  [key: string]: "available" | "occupied";
};

export interface BulkUpdateStatusRequest {
  statuses: BulkUpdateStatusRequestStatuses;
}

export type ListBookingsParams = {
  status?: ListBookingsStatus;
  date?: string;
};

export type ListBookingsStatus =
  (typeof ListBookingsStatus)[keyof typeof ListBookingsStatus];

export const ListBookingsStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
} as const;

export type BulkUpdateTableStatuses200 = {
  updated: number;
};