# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: bookings
    description: Table reservation operations
  - name: tables
    description: Table status operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a table reservation
      description: Creates a reservation using a PostgreSQL advisory lock to prevent concurrent double-bookings. A table with any active future reservation is fully locked until staff unlocks it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Reservation created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "409":
          description: Table already reserved or occupied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingError"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingError"
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all reservations
      description: Returns all reservations ordered by date and time, for staff use.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, confirmed, cancelled]
        - name: date
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"

  /bookings/{id}:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update reservation status
      description: Staff can confirm or cancel a pending reservation.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingStatusRequest"
      responses:
        "200":
          description: Reservation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Reservation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingError"

  /tables/statuses:
    get:
      operationId: getTableStatuses
      tags: [tables]
      summary: Get all table statuses
      description: Returns computed status for every table. Priority is staff override (occupied) > active booking (reserved) > available. Polls every few seconds for real-time updates.
      responses:
        "200":
          description: Map of table ID to status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableStatusMap"

  /tables/statuses/bulk:
    post:
      operationId: bulkUpdateTableStatuses
      tags: [tables]
      summary: Bulk update table statuses (staff only)
      description: Sets multiple tables to available or occupied at once. Setting a table to available also cancels its active future reservations.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkUpdateStatusRequest"
      responses:
        "200":
          description: Number of tables updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: integer
                required: [updated]

  /tables/{id}/status:
    patch:
      operationId: updateTableStatus
      tags: [tables]
      summary: Update a single table status (staff only)
      description: Staff sets a table to available or occupied. Setting to available also cancels active future reservations for that table.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTableStatusRequest"
      responses:
        "200":
          description: Updated table status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableStatusEntry"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingError"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    CreateBookingRequest:
      type: object
      properties:
        tableId:
          type: string
        hallName:
          type: string
        tableName:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        reservationDate:
          type: string
          description: Date in YYYY-MM-DD format
        reservationTime:
          type: string
          description: Time slot e.g. "07:00 PM"
        guestCount:
          type: integer
          minimum: 1
        specialRequest:
          type: string
          nullable: true
      required:
        - tableId
        - hallName
        - tableName
        - customerName
        - customerPhone
        - reservationDate
        - reservationTime
        - guestCount

    Booking:
      type: object
      properties:
        id:
          type: integer
        tableId:
          type: string
        hallName:
          type: string
        tableName:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        reservationDate:
          type: string
        reservationTime:
          type: string
        guestCount:
          type: integer
        specialRequest:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tableId
        - hallName
        - tableName
        - customerName
        - customerPhone
        - reservationDate
        - reservationTime
        - guestCount
        - status
        - createdAt

    UpdateBookingStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, cancelled]
      required:
        - status

    BookingError:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
      required:
        - error

    TableStatusMap:
      type: object
      additionalProperties:
        type: string
        enum: [available, reserved, occupied]

    TableStatusEntry:
      type: object
      properties:
        tableId:
          type: string
        status:
          type: string
          enum: [available, occupied]
      required:
        - tableId
        - status

    UpdateTableStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [available, occupied]
      required:
        - status

    BulkUpdateStatusRequest:
      type: object
      properties:
        statuses:
          type: object
          additionalProperties:
            type: string
            enum: [available, occupied]
      required:
        - statuses