# lib/api-zod/src/generated/types/tableStatusEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TableStatusEntryStatus } from "./tableStatusEntryStatus";

export interface TableStatusEntry {
  tableId: string;
  status: TableStatusEntryStatus;
}