# lib/api-zod/src/generated/types/tableStatusMap.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TableStatusMap {
  [key: string]: "available" | "reserved" | "occupied";
}