# lib/api-zod/src/generated/types/updateTableStatusRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTableStatusRequestStatus } from "./updateTableStatusRequestStatus";

export interface UpdateTableStatusRequest {
  status: UpdateTableStatusRequestStatus;
}